Add form inputs and description text
A row wrap was necessary to divide the form in two columns. Since a third 20vw input doesn't fit into the alocated space for the form section, it will start a new line below, and so on. A 20px margin-left was applied to every even input to add some space between then and the former one.