Style pseudoclasses, validate both password inputs
When invalid, inputs will have a 1px red outline. When valid, password inputs will gain a green outline and a check icon below them. The --content custom property was used to achieve this in javascript, defining a blank value for it in css, and with setProperty in javascript it is updated with the check icon. Its absolute positioning is relative to the <li> element, so we can put it at the <li> bottom right instead of inside the input.

When the value for the confirm-password input does not match its twin, the Create Account button is disabled via javascript.